voter_guide/requirements.txt
Django==1.10.2

# db related
psycopg2==2.6.2

# api related
djangorestframework==3.4.7
django-filter==0.15.2

# deploy related
gunicorn==19.6.0

# testing related
coverage==4.2

# others
django-pagination==1.0.7
django-debug-toolbar==1.6
